Move deploy functionality to a separate file
