+++ src/main/scala/com/lightbend/rp/sbtreactiveapp/App.scala
@@ -47,7 +47,7 @@ object App {
       .toLowerCase
 }
 
-sealed trait App extends SbtReactiveAppKeys {
+trait App extends SbtReactiveAppKeys {
   private def libIsPublished(scalaVersion: String) =
     SemVer
       .parse(scalaVersion)
@@ -188,10 +188,7 @@ case object PlayApp extends App {
       })
 }
 
-case object BasicApp extends App {
-  private val installReactiveSandbox = new java.util.concurrent.atomic.AtomicBoolean(false)
-  private val reactiveSandboxInstalledLatch = new java.util.concurrent.CountDownLatch(1)
-
+case object BasicApp extends DeployableApp {
   def globalSettings: Seq[Setting[_]] = Seq(
     annotations := Map.empty)
 
@@ -217,176 +214,14 @@ case object BasicApp extends App {
       aggregate in kubectl := false,
       aggregate in minikube := false)
 
-  def projectSettings: Seq[Setting[_]] =
-    Vector(
+  override def projectSettings: Seq[Setting[_]] =
+    super.projectSettings ++ Vector(
       alpinePackages := Vector.empty,
       appName := name.value,
       appType := "basic",
       applications := Vector("default" -> Vector(s"bin/${executableScriptName.value}")),
       cpu := 0.0D,
       diskSpace := 0L,
-      deployMinikubeEnableReactiveSandbox := {
-        val kafkaEnabled = SettingKey[Boolean]("lagomKafkaEnabled").?.value.getOrElse(false)
-        val cassandraEnabled = SettingKey[Boolean]("lagomCassandraEnabled").?.value.getOrElse(false)
-
-        kafkaEnabled || cassandraEnabled
-      },
-      deployMinikubeReactiveSandboxExternalServices := Map(
-        "cas_native" -> "_cql._tcp.reactive-sandbox-cassandra.default.svc.cluster.local",
-        "kafka_native" -> "_broker._tcp.reactive-sandbox-kafka.default.svc.cluster.local",
-        "elastic-search" -> "_http._tcp.reactive-sandbox-elasticsearch.default.svc.cluster.local"),
-      deployMinikubeAdditionalExternalServices := Map.empty,
-      deployMinikubeAkkaClusterBootstrapContactPoints := 1,
-      deployMinikubePlayHostAllowedProperty := "play.filters.hosts.allowed.0",
-      deployMinikubePlayHttpSecretKeyProperty := "play.http.secret.key",
-      deployMinikubePlayHttpSecretKeyValue := "dev-minikube",
-      deploy := {
-        import complete.DefaultParsers._
-        import scala.sys.process._
-
-        val args = spaceDelimited("<arg>").parsed
-        val isPlagom = Set("play", "lagom").contains(appType.value)
-        val bootstrapEnabled = enableAkkaClusterBootstrap.value
-        val reactiveSandbox = deployMinikubeEnableReactiveSandbox.value
-
-        args.headOption.getOrElse("").trim.toLowerCase match {
-          case "minikube" =>
-            // @TODO Windows support is partially implemented. When finishing impl, remove this guard.
-            // Issue that remains is that when arguments for rp have spaces, nodejs blows up
-
-            if (isWindows) {
-              sys.error("deploy is not currently supported on Microsoft Windows")
-            }
-
-            val minikubeExec =
-              if (isWindows)
-                target.value / "minikube-exec.ps1"
-              else
-                target.value / "minikube-exec"
-
-            val log = streams.value.log
-            val waitTimeMs = 1000 * 60 * 5
-
-            cmd.minikube.assert()
-            cmd.kubectl.assert()
-            cmd.rp.assert()
-
-            if (reactiveSandbox) {
-              cmd.helm.assert();
-            }
-
-            // This wrapper script that sets minikube environment before execing its args
-            // While it would be nice to do this all via the JVM, we need this mostly for hooking into
-            // the sbt-native-packager building.
-
-            IO.write(
-              minikubeExec,
-
-              if (isWindows)
-                """|minikube docker-env | Invoke-Expression
-                   |
-                   |$cmd, $as = $args
-                   |
-                   |& $cmd $as
-                   |""".stripMargin
-              else
-                """|#!/usr/bin/env bash
-                   |
-                   |set -e
-                   |
-                   |eval $(minikube docker-env --shell bash)
-                   |
-                   |exec "$@"
-                   |""".stripMargin)
-
-            assert(minikubeExec.setExecutable(true), s"Failed to mark $minikubeExec as executable")
-
-            // We install the sandbox now (in on task via AtomicBoolean) but don't wait until after the build is
-            // done for it to be deployed. This saves a bit of time for the user.
-
-            val shouldInstallReactiveSandbox = reactiveSandbox && installReactiveSandbox.compareAndSet(false, true)
-
-            if (shouldInstallReactiveSandbox) {
-              if (!cmd.kubectl.deploymentExists("kube-system", "tiller-deploy")) {
-                cmd.helm.init(log)
-
-                cmd.kubectl.waitForDeployment(log, "kube-system", "tiller-deploy", waitTimeMs = waitTimeMs)
-              }
-
-              if (!cmd.kubectl.deploymentExists("default", "reactive-sandbox")) {
-                cmd.helm.installReactiveSandbox(log)
-              }
-            }
-
-            val minikubeIp = cmd.minikube.ip()
-
-            val javaOpts =
-              Vector(
-                if (isPlagom) s"-D${deployMinikubePlayHostAllowedProperty.value}=$minikubeIp" else "",
-                if (isPlagom) s"-D${deployMinikubePlayHttpSecretKeyProperty.value}=${deployMinikubePlayHttpSecretKeyValue.value}" else "")
-                .filterNot(_.isEmpty)
-
-            val services =
-              if (reactiveSandbox)
-                deployMinikubeReactiveSandboxExternalServices.value ++ deployMinikubeAdditionalExternalServices.value
-              else
-                deployMinikubeAdditionalExternalServices.value
-
-            val serviceArgs =
-              services.flatMap {
-                case (serviceName, serviceAddress) =>
-                  Vector("--external-service", s"$serviceName=$serviceAddress")
-              }
-
-            val rpArgs =
-              Vector(
-                dockerAlias.value.versioned,
-                "--env",
-                s"JAVA_OPTS=${javaOpts.mkString(" ")}") ++
-                (if (bootstrapEnabled) Vector("--akka-cluster-skip-validation", "--pod-controller-replicas", deployMinikubeAkkaClusterBootstrapContactPoints.value.toString) else Vector.empty) ++
-                serviceArgs ++
-                deployMinikubeRpArguments.value
-
-            publishLocalDocker(
-              (stage in Docker).value,
-              if (isWindows)
-                "powershell.exe" +: minikubeExec.getAbsolutePath +: dockerBuildCommand.value
-              else
-                minikubeExec.getAbsolutePath +: dockerBuildCommand.value,
-              log)
-
-            log.info(s"Built image ${dockerAlias.value.versioned}")
-
-            if (reactiveSandbox) {
-              // FIXME: Make tiller & reactive-sandbox names configurable
-
-              cmd.kubectl.waitForDeployment(log, "default", "reactive-sandbox", waitTimeMs = waitTimeMs)
-
-              if (shouldInstallReactiveSandbox) {
-                for {
-                  pod <- cmd.kubectl.getPodNames("app=reactive-sandbox")
-                  statement <- (deployMinikubeReactiveSandboxCqlStatements in ThisBuild).value
-                } {
-                  log.info(s"executing cassandra cql: $statement")
-
-                  cmd.kubectl.invoke(log, Seq("exec", pod, "--", "/bin/bash", "-c", s"""/opt/cassandra/bin/cqlsh "$$POD_IP" -e "$statement""""))
-                }
-
-                reactiveSandboxInstalledLatch.countDown()
-              } else {
-                reactiveSandboxInstalledLatch.await()
-              }
-            }
-
-            val kubernetesResourcesYaml = cmd.rp.generateKubernetesResources(minikubeExec.getAbsolutePath, log, rpArgs)
-
-            cmd.kubectl.deleteAndApply(log, kubernetesResourcesYaml)
-
-          case other =>
-            sys.error(s"""Unknown deployment target: "$other". Available: minikube""")
-        }
-      },
-      deployMinikubeRpArguments := Seq.empty,
       memory := 0L,
       enableCGroupMemoryLimit := true,
       privileged := false,
